project_p2q3e_main/src/main/persistence/JsonWriter.java
package persistence;

import model.Connect4Game;
import org.json.JSONObject;

import java.io.File;
import java.io.FileNotFoundException;
import java.io.PrintWriter;
import model.Event;
import model.EventLog;

// used 210's JsonSerializationDemo as a model for this class
public class JsonWriter {
    private static final int TAB = 4;
    private PrintWriter writer;
    private String destination;
    private EventLog log;
    // EFFECTS: constructs writer to write to destination file

    public JsonWriter(String destination) {
        log = EventLog.getInstance();
        this.destination = destination;
    }

    // MODIFIES: this
    // EFFECTS: opens writer; throws FileNotFoundException if destination file cannot
    // be opened for writing
    public void open() throws FileNotFoundException {
        writer = new PrintWriter(new File(destination));
    }

    // MODIFIES: this
    // EFFECTS: writes JSON representation of Connect4Game to file
    public void write(Connect4Game game) {
        JSONObject json = game.toJson();
        saveToFile(json.toString(TAB));
    }

    // MODIFIES: this
    // EFFECTS: closes writer
    public void close() {
        writer.close();
    }

    // MODIFIES: this
    // EFFECTS: writes string to file
    private void saveToFile(String json) {
        writer.print(json);
        log.logEvent(new Event("File has been saved Successfully"));
    }
}
